+++ packages/beej-react/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @beej/react
 
+## 0.0.9
+
+### Patch Changes
+
+- 6e52cdf: formatting fixes and package name update
+- 47300d1: fix: package json dependencies correct list
+
+## 0.0.8
+
+### Patch Changes
+
+- fix template not published issue
+
 ## 0.0.7
 
 ### Patch Changes